PHP Debug is WORKING!
Thanks to a sample from the kickstarter repo, I was able to get the
basic debug adapter in place.  Grumpy Learning got me the rest of the
way with getting the syntax in place for the PHP adapter.

It is really just so beautiful.  Still, likely need a few more key
mapping additions and changes to fit my workflow.