Replaced intelphense with phpactor
Mostly seems to be working.  Only one issue, and that was error message
for missing docblock parameters.  This was corrected by creating a file
called ~/.config/phpactor/phpactor.yml and putting this into it.

{
    "worse.docblock_missing_param": false
}

And now it seems to be working.  Going to ride with this for a while to
see how it compares.