Update how I call flash
Going to try this for a little while to see if this flows better for me.
The leader-j approach just feels too much out of the way.  Likely going
to tweak on this one again until it feels right.

Also disabled the toggle.  Not sure what it was supposed to do, but it
didn't do it for me.